updates website
I've been working all morning trying to get a structure to the entire website. I've been wokring on the form for people to sign up which is pretty pleasing to the eye now (add any other info we should ask for)! The links should all work now to take you between pages, including the links for the contact forms (css still needs to be adjusted)